dataset/cwe_89_sql_injection/testcase6_secure_second_layer/requirements.txt
SQLAlchemy==2.0.29
Flask-SQLAlchemy==3.1.0
